Update class modal
