+++ src/pages/classes/ClassModal/Steps/Fourth/class.modal.steps.fourth.interface.ts
@@ -1,23 +1,22 @@
-import { ClassResponse } from 'models/http/responses/class.response.models';
-import { ClassModalStepsProps } from '../class.modal.steps.interface';
-import { ClassFirstForm } from '../First/class.modal.steps.first.interface';
-import { ClassSecondForm } from '../Second/class.modal.steps.second.interface';
-import { ClassThirdForm } from '../Third/class.modal.steps.third.interface';
 import { SubjectResponse } from 'models/http/responses/subject.response.models';
-import { CalendarResponse } from 'models/http/responses/calendar.responde.models';
-import { ScheduleData } from '../../class.modal.interface';
+import { ClassModalStepsProps } from '../class.modal.steps.interface';
+import { UseFormReturn } from 'react-hook-form';
 
-interface ClassFormData {
-  first: ClassFirstForm,
-  second: ClassSecondForm,
-  third: ClassThirdForm
-}
 export interface ClassModalFourthStepProps extends ClassModalStepsProps {
-  data: ClassFormData,
-  onNext: () => void;
-  moveTo: (index: number) => void;
+  form: UseFormReturn<ClassFourthForm, any, ClassFourthForm>;
   subjects: SubjectResponse[];
-  calendars: CalendarResponse[];
-  schedules: ScheduleData[];
-  selectedClass?: ClassResponse;
+  onNext: (data: ClassFourthForm) => void;
 }
+export interface ClassFourthForm {
+  ignore_to_allocate: boolean;
+  projector: boolean;
+  air_conditionating: boolean;
+  accessibility: boolean;
+}
+
+export type fieldNames =
+  | 'ignore_to_allocate'
+  | 'projector'
+  | 'air_conditionating'
+  | 'accessibility'
+  | 'skip_exceptions';
