+++ src/pages/classes/ClassModal/Steps/Third/class.modal.steps.third.interface.ts
@@ -1,22 +1,34 @@
-import { SubjectResponse } from 'models/http/responses/subject.response.models';
 import { ClassModalStepsProps } from '../class.modal.steps.interface';
 import { UseFormReturn } from 'react-hook-form';
+import { WeekDay } from 'utils/enums/weekDays.enum';
+import { MonthWeek } from 'utils/enums/monthWeek.enum';
+import { Recurrence } from 'utils/enums/recurrence.enum';
+import { ScheduleData } from '../../class.modal.interface';
+import { ClassSecondForm } from '../Second/class.modal.steps.second.interface';
 
 export interface ClassModalThirdStepProps extends ClassModalStepsProps {
   form: UseFormReturn<ClassThirdForm, any, ClassThirdForm>;
-  subjects: SubjectResponse[];
-  onNext: (data: ClassThirdForm) => void;
+  secondForm: UseFormReturn<ClassSecondForm, any, ClassSecondForm>;
+  schedules: ScheduleData[];
+  setSchedules: (schedules: ScheduleData[]) => void;
+  onNext: () => void;
 }
 export interface ClassThirdForm {
-  ignore_to_allocate: boolean;
-  projector: boolean;
-  air_conditionating: boolean;
-  accessibility: boolean;
+  start_time?: string;
+  end_time?: string;
+  week_day?: WeekDay;
+  month_week?: MonthWeek;
+  recurrence?: Recurrence;
+  schedule_start_date?: string;
+  schedule_end_date?: string;
 }
 
+export type numberRange = 0 | 1 | 2 | 3 | 4 | 5 | 6;
 export type fieldNames =
-  | 'ignore_to_allocate'
-  | 'projector'
-  | 'air_conditionating'
-  | 'accessibility'
-  | 'skip_exceptions';
+  | 'recurrence'
+  | 'week_day'
+  | 'month_week'
+  | 'schedule_start_date'
+  | 'schedule_end_date'
+  | 'start_time'
+  | 'end_time';
