+++ src/utils/schedules/schedules.validator.ts
@@ -1,11 +1,16 @@
 import moment from 'moment';
 import CommonValidator from 'utils/common/common.validator';
+import { MonthWeek } from 'utils/enums/monthWeek.enum';
 import { Recurrence } from 'utils/enums/recurrence.enum';
 import { WeekDay } from 'utils/enums/weekDays.enum';
 
 export class ScheduleValidator extends CommonValidator {
   static isInvalidWeekDay(value: number) {
     return !WeekDay.getValues().includes(value as WeekDay);
+  }
+
+  static isInvalidMonthWeek(value: number) {
+    return !MonthWeek.getValues().includes(value as MonthWeek);
   }
 
   static isInvalidDayTime(value: string) {
